Move filter structure to page component
