+++ src/attributes/components/AttributeListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./AttributeListPage";
 export * from "./AttributeListPage";
+export * from "./filters";
